Move listing to osfHelpers, use VM map function for listing
