+++ website/static/js/logFeed.js
@@ -10,55 +10,10 @@ var Paginator = require('js/paginator');
 var oop = require('js/oop');
 require('knockout.punches');
 
-var $osf = require('./osfHelpers');
+var $osf = require('js/osfHelpers');  // Injects 'listing' binding hanlder to to Knockout
+var nodeCategories = require('json!built/nodeCategories.json');
 
 ko.punches.enableAll();  // Enable knockout punches
-
-
-/* A binding handler to convert lists into formatted lists, e.g.:
- * [dog] -> dog
- * [dog, cat] -> dog and cat
- * [dog, cat, fish] -> dog, cat, and fish
- *
- * This handler should not be used for user inputs.
- */
-ko.bindingHandlers.listing = {
-    update: function(element, valueAccessor, allBindings, viewModel, bindingContext) {
-        var value = valueAccessor();
-        var valueUnwrapped = ko.unwrap(value);       
-        var map = valueUnwrapped.map || function(item) {return item;};
-        var data = valueUnwrapped.data || [];
-        var keys = [];
-        if (!Array.isArray(data)) {
-            keys = Object.keys(data);
-        }
-        else {
-            keys = data;
-        }
-        var index = 1;
-        var list = ko.utils.arrayMap(keys, function(key) {
-            var ret;
-            if (index === 1){
-                ret = '';
-            }
-            else if (index === 2){
-                if (valueUnwrapped.length === 2) {
-                    ret = ' and ';
-                }
-                else {
-                    ret = ', ';
-                }
-            }
-            else {
-                ret = ', and ';
-            }
-            ret += map(key, data[key]);
-            index++;
-            return ret;
-        }).join('');
-        $(element).html(list);
-    }
-};
 
 /**
   * Log model.
@@ -89,6 +44,14 @@ var Log = function(params) {
         return $('script#' + self.action).length > 0;
     });
 
+    self.mapUpdates = function(key, item) {
+        if (key === 'category') {
+            return key + ' to ' + nodeCategories[item['new']];
+        }
+        else {
+            return key + ' to ' + item;
+        }
+    };
 
     /**
       * Return the html for a comma-delimited list of contributor links, formatted
